+++ include/tools/types/__void_t/void_t-2015.hpp
@@ -1,0 +1,8 @@
+// [2021y-04m-13d][15:09:41] Idrisov Denis R. 100 PRE
+#pragma once
+#ifndef dTOOLS_TYPES_VOID_T_USED_ 
+#define dTOOLS_TYPES_VOID_T_USED_ 100 PRE
+
+//==============================================================================
+//==============================================================================
+#endif // !dTOOLS_TYPES_VOID_T_USED_