+++ include/tools/types/common/is_dereferencable.hpp
@@ -1,8 +1,9 @@
 // [2021y-04m-06d][20:54:59] Idrisov Denis R. 100
 // [2021y-04m-08d][01:38:17] Idrisov Denis R. 101
+// [2021y-04m-13d][15:05:11] Idrisov Denis R. 102 PRE
 #pragma once
 #ifndef dTOOLS_IS_DEREFERENCABLE_USED_ 
-#define dTOOLS_IS_DEREFERENCABLE_USED_ 101
+#define dTOOLS_IS_DEREFERENCABLE_USED_ 102 PRE
 
 #include <tools/features.hpp>