+++ include/tools/types/common/is_dereferencable/is_dereferencable-2013.hpp
@@ -1,8 +1,9 @@
 // [2021y-04m-06d][21:36:00] Idrisov Denis R.
 // [2021y-04m-08d][01:38:17] Idrisov Denis R. 101
+// [2021y-04m-13d][15:05:41] Idrisov Denis R. 102 PRE
 #pragma once
 #ifndef dTOOLS_IS_DEREFERENCABLE_2013_USED_ 
-#define dTOOLS_IS_DEREFERENCABLE_2013_USED_ 101,2013
+#define dTOOLS_IS_DEREFERENCABLE_2013_USED_ 102,2013
 
 #include <tools/features.hpp>
 #include <type_traits>
@@ -13,6 +14,8 @@ namespace tools
 {
     namespace detail
     {
+        template<class t> t val();
+
         template<class t, bool> class is_deref_
         {
             using x = ::std::remove_reference_t<t>;
@@ -21,7 +24,7 @@ namespace tools
                 decltype(__VA_ARGS__, ::std::true_type{})
 
             template<class u> static dCHECK_EXPRESSION( 
-                *::std::declval<u>()
+                *val<u&>()
             ) check(u*);
 
             #undef dCHECK_EXPRESSION